srv: implement monitor-data cap'ing
This CL makes sure the amount of memory used for monitoring is under
control.
When the slice used to keep and display monitoring data grows to 512,
it's trimmed to 256 (keeping one element every 2 elements.)
Also, when the first data point is older than 6h, reset the slice to a
length of 1, only keeping the most recent data point.

Fixes #11.